Initial DXTV intra-frame compression

+++ src/codec_dxtv.h
@@ -11,7 +11,7 @@ public:
     /// Differences:
     /// - Colors will be stored as RGB555 only
     /// - Blocks are stored sequentially from left to right, top to bottom, but colors and indices are stored separately. First all colors, then all indices
-    static auto encodeDXTV(const std::vector<uint16_t> &image, uint32_t width, uint32_t height, bool keyFrame, float maxBlockError = 1.0F) -> std::vector<uint8_t>;
+    static auto encodeDXTV(const std::vector<uint16_t> &image, uint32_t width, uint32_t height, bool keyFrame, double maxBlockError = 1.0) -> std::vector<uint8_t>;
 
     /// @brief Decompress from DXTV format.
     static auto decodeDXTV(const std::vector<uint8_t> &data, uint32_t width, uint32_t height) -> std::vector<uint8_t>;
